Separate out Site Health code
